Move logic to Interactions to clean Card up even more
