+++ front/app/components/IdeaCard/Interactions.tsx
@@ -1,24 +1,32 @@
 import React from 'react';
 
+// api
+import useBasket from 'api/baskets/useBasket';
+
 // config
 import { getVotingMethodConfig } from 'utils/configs/votingMethodConfig';
 
 // utils
 import { isCurrentPhase } from 'api/phases/utils';
+import { pastPresentOrFuture } from 'utils/dateUtils';
 
 // types
 import { IIdea } from 'api/ideas/types';
 import { IProjectData } from 'api/projects/types';
 import { IPhaseData } from 'api/phases/types';
 
-type InteractionsProps = {
+type Props = {
   idea: IIdea;
   participationContext?: IPhaseData | IProjectData | null;
 };
 
-const Interactions = ({ participationContext, idea }: InteractionsProps) => {
+const Interactions = ({ participationContext, idea }: Props) => {
+  const isGeneralIdeasPage = window.location.pathname.endsWith('/ideas');
   const votingMethod = participationContext?.attributes.voting_method;
   const config = getVotingMethodConfig(votingMethod);
+  const { data: basket } = useBasket(
+    participationContext?.relationships?.user_basket?.data?.id
+  );
 
   if (!config || !participationContext) return null;
 
@@ -26,6 +34,20 @@ const Interactions = ({ participationContext, idea }: InteractionsProps) => {
     participationContext.type === 'phase' &&
     !isCurrentPhase(participationContext)
   ) {
+    return null;
+  }
+
+  const participationContextEnded =
+    participationContext.type === 'phase' &&
+    pastPresentOrFuture(participationContext.attributes.end_at) === 'past';
+
+  const hideInteractions =
+    isGeneralIdeasPage ||
+    (participationContextEnded && basket?.data.attributes.submitted_at === null)
+      ? true
+      : false;
+
+  if (hideInteractions) {
     return null;
   }
